rmove step

+++ Jenkinsfile
@@ -146,16 +146,19 @@ pipeline {
                 script {
                     sh 'whoami'
                     sh 'pwd'
-                    sh '''
-                            ssh -T -i ${KEY_PATH} -o StrictHostKeyChecking=no ${SSH_USER}@${SERVER_IP} << EOF
+                    sh """
+                            ssh -T -i \$SSH_KEY -o StrictHostKeyChecking=no ${SSH_USER}@${SERVER_IP} << EOF
                             docker login -u $DOCKER_USERNAME -p $DOCKER_PASSWORD ${REGISTRY_URL}
-                            docker pull ${REGISTRY_URL}/${DOCKER_IMAGE}:${TAG_IMAGE}
-                            docker stop $(docker ps -a -q) || true
-                            docker rm $(docker ps -a -q) || true
-                            docker run -d -p 3000:3000 --name nestjs-docker ${REGISTRY_URL}/${DOCKER_IMAGE}:${TAG_IMAGE}
-                            docker image prune -f
+                            if docker pull ${REGISTRY_URL}/${DOCKER_IMAGE}:${TAG_IMAGE}; then
+                                docker stop \$(docker ps -a -q) || true
+                                docker rm \$(docker ps -a -q) || true
+                                docker run -d -p 3000:3000 --name nestjs-docker ${REGISTRY_URL}/${DOCKER_IMAGE}:${TAG_IMAGE}
+                                docker image prune -f
+                            else
+                                echo "Failed to pull the Docker image"
+                            fi
                             EOF
-                        '''
+                        """
                     // withCredentials([sshUserPrivateKey(credentialsId: SSH_CREDENTIALS_ID, keyFileVariable: 'SSH_KEY')]) {
                     //     // ssh -i $SSH_KEY -o StrictHostKeyChecking=no ${SSH_USER}@${SERVER_IP} << EOF
                     //     sh '''
@@ -178,18 +181,17 @@ pipeline {
                 }
             }
         }
+    }
 
-        stage('Cleanup') {
-            steps {
-                sh '''
-                    # Docker 이미지 및 컨테이너 정리
-                    docker system prune -a -f --volumes
-
-                    # 임시 파일 및 캐시 정리
-                    rm -rf /tmp/*
-                    rm -rf /var/cache/apk/*
-                '''
-            }
+    post {
+        always {
+            sh '''
+                echo "Final cleanup..."
+                docker system prune -a -f --volumes
+                rm -rf /tmp/*
+                rm -rf /var/cache/apk/*
+                echo "Pipeline completed"
+            '''
         }
     }
 }
